pneumonia-detector/requirements.txt
torch==2.9.0
torchvision==0.24.0
fastapi==0.110.0
uvicorn[standard]==0.27.0
python-multipart==0.0.9
pillow>=11.0.0
scikit-learn==1.4.0
matplotlib==3.8.0
numpy==1.26.0
opencv-python-headless==4.9.0.80
kaggle==1.6.6
jupyter==1.0.0
ipykernel==6.29.0